// src/app/pages/home/types/home.type.ts
export interface Product {
  id?: string;
  code?: string;
  name?: string;
  description?: string;
  price?: number;
  quantity?: number;
  inventoryStatus?: string;
  category?: string;
  image?: string;
  rating?: number;
}

export interface OptionLocation {
  center: { lat: number; lng: number };
  zoom: number;
}